api/Vexora-Model/requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
numpy
python-multipart
python-dotenv